+++ CHANGELOG.md
@@ -3,6 +3,12 @@
 All notable changes to this project will be documented in this file.
 This project uses [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Added
+
+- Types, where possible, now implement `Ord`, `PartialOrd` and `Hash`.
+
 ## [0.2.3] - 2024-07-03
 
 [0.2.3]: https://github.com/sunsided/serial-sensors-proto/releases/tag/v0.2.3